requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==1.4.46
pymysql==1.0.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
bcrypt==4.0.1
boto3==1.26.137
pydantic==1.10.7
email-validator==2.0.0.post2
pytest==7.3.1
requests==2.30.0